rename ds package to pstoreds.
